requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0

# AI and ML
openai>=1.0.0
anthropic>=0.3.0

# Financial data
yfinance>=0.2.0
alpha-vantage>=2.3.0

# Data visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Presentation generation
python-pptx>=0.6.21
Pillow>=9.0.0

# Web interface
streamlit>=1.25.0
gradio>=3.40.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
click>=8.0.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
